Added influxdb, itemmeta conversion, scheduler abstraction

+++ SkyblockCore-Common/src/main/java/me/illusion/skyblockcore/common/utilities/time/Time.java
@@ -21,7 +21,7 @@ public class Time {
      * @param unit The unit to convert to
      * @return The converted time
      */
-    public long as(TimeUnit unit) {
+    public int as(TimeUnit unit) {
         return unit.convert(duration, this.unit);
     }
